Improvements of Modules lesson according to feedback
