+++ Modules and packages/Built-in modules/task.md
@@ -20,7 +20,7 @@ secondary prompts if the interpreter is in the interactive mode:
 The variable `sys.path` is a list of strings that determines the interpreter’s search path 
 for modules: see what it prints for you when you run the code of the task.
 
-Remember that you can use &shortcut:CodeCompletion; after a dot (.) to explore available 
+Remember that you can use &shortcut:CodeCompletion; shortcut after a dot (.) to explore available 
 methods of a module. You can read more about standard modules <a href="https://docs.python.org/3/tutorial/modules.html#standard-modules">here</a>.
 
 For more structured and detailed information, you can also refer to [this Hyperskill knowledge base page](https://hyperskill.org/learn/step/6019#built-in-modules).
