+++ Modules and packages/Import module 2/imports.py
@@ -1,0 +1,11 @@
+import calculator
+
+calc = calculator.Calculator()
+for i in range(100):
+    calc.add(i)
+
+print(calc.get_current())
+
+
+
+
