+++ Modules and packages/Import module 2/task.md
@@ -1,0 +1,15 @@
+## Import module 2
+
+You can import not only functions: you can import classes or even other modules. It is customary but not required to place all 
+import statements at the beginning of a module.
+
+You can find out more about modules in Python by reading [this section](https://docs.python.org/3/tutorial/modules.html) of The Python Tutorial.
+
+### Task
+In the code editor, import the module `calculator` and create an instance of the class `Calculator` (`calc`). 
+Use the `add` method defined in `Calculator` in a loop to add up numbers from 0 to 99.
+
+<div class='hint'>Use the <code>import</code> keyword and the <code>calculator</code> reference.</div>
+<div class='hint'>Access the function from the module using syntax such as <code>module.function()</code>.</div>
+<div class="hint">Don't forget to provide the function with an argument.</div>
+
