+++ Modules and packages/Import module/imports.py
@@ -1,15 +1,3 @@
-import my_module
-import calculator
+import my_funcs
 
-my_module.hello_world("John")
-
-
-calc = calculator.Calculator()
-for i in range(100):
-    calc.add(i)
-
-print(calc.get_current())
-
-
-
-
+my_funcs.hello_world("John")
