+++ Modules and packages/lesson-info.yaml
@@ -1,6 +1,7 @@
 content:
-- Import module
-- Built-in modules
-- From import
-- Packages
-- Executing modules as scripts
+  - Import module
+  - Import module 2
+  - Built-in modules
+  - From import
+  - Packages
+  - Executing modules as scripts
